requirements.txt
deebot-client>=13.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.8.0